Minor adjustemetns
